pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic<=2.12.5",
    "aiogram<=3.26.0",
    "python-dotenv<=1.2.2",
    "requests<=2.32.5",
    "python-dateutil<=2.9.0.post0",
]
